requirements.txt
Django~=5.1.3
asgiref~=3.8.1
cffi~=1.17.1
cryptography~=43.0.3
django-encrypted-model-fields~=0.6.5
django-environ~=0.11.2
django-fernet-fields~=0.6
django-phonenumber-field~=8.0.0
phonenumbers~=8.13.49
psycopg2-binary~=2.9.10
pycparser~=2.22
sqlparse~=0.5.2
typing-extensions~=4.12.2
tzdata~=2024.2
wheel~=0.45.0
djangorestframework~=3.15.2